chore: Update `CheckoutCalculateTaxResponse` with new fields
