+++ src/stripe/stripe-service.ts
@@ -8,7 +8,7 @@ import { loadStripe } from "@stripe/stripe-js";
 import type { BrandingInfoResponse } from "../networking/responses/branding-response";
 import { Theme } from "../ui/theme/theme";
 import { DEFAULT_TEXT_STYLES } from "../ui/theme/text";
-import type { GatewayParams } from "../networking/responses/checkout-start-response";
+import type { GatewayParams } from "../networking/responses/stripe-elements";
 export class StripeService {
   private static FORM_VALIDATED_CARD_ERROR_CODES = [
     "card_declined",
